Soft Body Demos - Pressure Spheres, Volume Cubes, new Auto Constraints (#148)
This adds in some of the Jolt 5.0 LRA soft body constrain code, using SoftBodySharedSettings_ELRAType

One of the files is riskier, with the example.js saving direct mapping of the MotionProperties pointers to vertex location. This was partly when I was trying to skin to other models, and also avoid rebuilding 16+ soft models via ShapeGetTriangles every frame. It is coded specifically to the SoftVertex CurrentPosition offset, based on PreviousPosition being before it and using 16 byte alignment, but should Jolt ever move the data location during runtime, or the build and class structure ever change, or any of this Soft Vertex usage of Vec3 ever change to RVec3, then this code would break or need to be updated. If WebIDL supported [Ref] attributes, this could at least be cleaner than hard-coded offsets into class structs. This file can be discarded if wanted.

The 5.0.0 SoftBodyCreator was ported in almost 1-to-1, with the only difference being Cube supporting compliance values in its factory, and perturbation does not return a Jolt memory bound Vec3 given manual memory management.

The remaining code is fairly 1-to-1 for the following JoltPhysics Samples:

* SoftBodyBendConstraintTest.cpp
* SoftBodyFrictionTest.cpp
* SoftBodyGravityFactorTest.cpp
* SoftBodyLRAConstraintTest.cpp
* SoftBodyPressureTest.cpp

Demo of Cube volume, although there is not easily a way to show the range of impact of Compliance in an X/Y chart without having one or two entries fly away. It is also hard to get good jelly colors (vibrant and transparent), so the demo can always be changed to just use regular solid colors with no transparent shaders.

Some of the demos seem a little redundant like Gravity and Friction, although it does show how soft bodies act different than rigid in those case like pressure having more impact in low-grav, or how a soft-sphere rolls by spilling over itself.